parseClueID works as intended for standard case and empty case.
Standardized filepath for QR locations defined as QR_PATH

string value webLink defined, however webLink requires more complex logic and thinking.  It is the value that is encoded in each QR Code.  Thus, it needs to perform specific logic when scanned by the phone app.  The phone app will handle communicating with the web app, but the QR Code needs to contain the request body or values that can be used to form the request.

Currently working on method genQRCode(int, int).  Resolve previous issues and work can easily continue.