insertClueIntoTable added, code cleanup
Cleaned up some code fore readability, making sure exceptions are handled at the lowest level possible.  

Implemented insertClueIntoTable, succeeds in base case.  Bug found in method and reported (#1).  

insertClueIntoTable takes "String fbid" as an argument, which is the Facebook Group ID of the Huntsperson group the clues belong to.  As this functionality is not implemented, this field contains a blank string.