Finished pursuing dead end, cleaned up dead end code.
Attempts to post the group from JS failed.  The problem arose that I couldn't authenticate the group request from the JS page because I didn't have the app secret, but practice forbids me from using the app secret client side due to security concerns.  An alternate solution presents itself in the form of server side HTTP requests to the facebook api, which can be used to create the group.  The request has to be made manually, but I was already trying to do that client side.