Added method and logic to create a facebook group in method makeGroup(HttpServletRequest).
This method is a pain in the ass.  The servlet will fire a POST to the facebook graph api to make a new group along the Apps and Games edge.  However, the access token is input manually, being hardcoded into the servlet.  If the app secret is reset then the method won't work.  It's horrible practice, but hastens the dev process.

Added a few new final strings to represent this app data.  

makeGroup will return a group id, suggesting the POST was successful.  However, it's encoded in a JSON object and I haven't developed the logic to extract the id or put it to good use.